// src/models/Collegio.ts
// Generated by https://quicktype.io
import {Deputato} from './Deputato'

export interface Collegio {
    gid: number
    objectid: string
    cod_reg: string
    regione: string
    cod_pro: string
    provincia: string
    cod_cm: number
    cittametro: string
    pro_com17: string
    comune17: string
    pro_com11: string
    comune11: string
    cod_asc_93: string
    asc_93: undefined
    pop_2011: string
    circo17_c: string
    circo17_d: string
    cod_ita_18: string
    cam17u_cod: string
    cam17u_den: string
    cam17u_nom: string
    cam17p_cod: string
    cam17p_den: string
    sen17u_cod: string
    sen17u_den: string
    sen17u_nom: string
    sen17p_cod: string
    sen17p_den: string
    shape_leng: string
    shape_area: string
    simple_geom: undefined
    estero: boolean
    stato: string
    continente: string
    deputati: Deputato[]
    geoJsonMap: string
}

export interface RisultatiPlurinominale {
    plurinominale: string
    lista: string
    sigla_lista: string
    coalizione: string
    sigla_coalizione: string
    voti: string
    percentuale: string
    voti_coalizione: string
    tipo: string
}

export interface RisultatiUninominale {
    nome: string
    cognome: string
    coalizione: string
    sigla_coalizione: string
    voti: string
    percentuale: string
    tipo: string
}